merge code

+++ src/pages/add/copayers/copayers.ts
@@ -7,8 +7,8 @@ import { Subscription } from 'rxjs';
 import { SocialSharing } from '@ionic-native/social-sharing';
 
 // Providers
-import { ActionSheetProvider } from '../../../providers/action-sheet/action-sheet';
 import { AppProvider } from '../../../providers/app/app';
+import { ConfigProvider } from '../../../providers/config/config';
 import { KeyProvider } from '../../../providers/key/key';
 import { Logger } from '../../../providers/logger/logger';
 import { OnGoingProcessProvider } from '../../../providers/on-going-process/on-going-process';
@@ -27,8 +27,10 @@ export class CopayersPage {
   public isCordova: boolean;
 
   public wallet;
+  public canSign: boolean;
   public copayers: any[];
   public secret;
+  public useLegacyQrCode: boolean;
 
   private onResumeSubscription: Subscription;
   private onPauseSubscription: Subscription;
@@ -48,8 +50,8 @@ export class CopayersPage {
     private translate: TranslateService,
     private pushNotificationsProvider: PushNotificationsProvider,
     private viewCtrl: ViewController,
-    private actionSheetProvider: ActionSheetProvider,
-    private keyProvider: KeyProvider
+    private keyProvider: KeyProvider,
+    private configProvider: ConfigProvider
   ) {
     this.secret = null;
     this.appName = this.appProvider.info.userVisibleName;
@@ -57,36 +59,31 @@ export class CopayersPage {
     this.isCordova = this.platformProvider.isCordova;
     this.copayers = [];
     this.wallet = this.profileProvider.getWallet(this.navParams.data.walletId);
+    this.canSign = this.wallet.canSign;
+    this.useLegacyQrCode = this.configProvider.get().legacyQrCode.show;
   }
 
   ionViewDidLoad() {
     this.logger.info('Loaded: CopayersPage');
+  }
 
+  ngOnInit() {
+    this.subscribeEvents();
+    this.events.publish('Local/WalletFocus', {
+      walletId: this.wallet.credentials.walletId
+    });
     this.onResumeSubscription = this.plt.resume.subscribe(() => {
       this.events.publish('Local/WalletFocus', {
         walletId: this.wallet.credentials.walletId
       });
       this.subscribeEvents();
     });
-
     this.onPauseSubscription = this.plt.pause.subscribe(() => {
       this.unsubscribeEvents();
     });
   }
 
-  ionViewWillEnter() {
-    this.events.publish('Local/WalletFocus', {
-      walletId: this.wallet.credentials.walletId
-    });
-    this.subscribeEvents();
-  }
-
-  ionViewWillLeave() {
-    this.unsubscribeEvents();
-  }
-
   ngOnDestroy() {
-    this.events.publish('Local/WalletListChange');
     this.onResumeSubscription.unsubscribe();
     this.onPauseSubscription.unsubscribe();
   }
@@ -100,6 +97,7 @@ export class CopayersPage {
   }
 
   close() {
+    this.unsubscribeEvents();
     this.viewCtrl.dismiss();
   }
 
@@ -117,30 +115,38 @@ export class CopayersPage {
         // TODO?
         this.wallet.openWallet(err => {
           if (err) this.logger.error(err);
-          this.viewCtrl.dismiss().then(() => {
-            this.events.publish('Local/WalletListChange');
-            this.events.publish('OpenWallet', this.wallet);
-          });
+          this.close();
         });
       }
     }
   }
 
   public showDeletePopup(): void {
-    let title = this.translate.instant('Confirm');
-    let msg = this.translate.instant(
-      'Are you sure you want to cancel and delete this wallet?'
-    );
+    const title = this.translate.instant('Confirm');
+    let msg;
+    if (!this.canSign) {
+      msg = this.translate.instant(
+        'Are you sure you want to delete this wallet?'
+      );
+    }
+    msg = this.translate.instant('Are you sure you want to hide this wallet?');
     this.popupProvider.ionicConfirm(title, msg).then(res => {
       if (res) this.deleteWallet();
     });
   }
 
   private deleteWallet(): void {
+    if (this.canSign) {
+      this.profileProvider.toggleHideWalletFlag(this.wallet.id);
+      setTimeout(() => {
+        this.close();
+      }, 1000);
+      return;
+    }
     this.onGoingProcessProvider.set('deletingWallet');
     this.profileProvider
       .deleteWalletClient(this.wallet)
-      .then(async () => {
+      .then(() => {
         this.onGoingProcessProvider.clear();
         this.pushNotificationsProvider.unsubscribe(this.wallet);
 
@@ -149,33 +155,20 @@ export class CopayersPage {
           const keyInUse = this.profileProvider.isKeyInUse(keyId);
 
           if (!keyInUse) {
-            await this.keyProvider.removeKey(keyId);
-            delete this.profileProvider.walletsGroups[keyId];
+            this.keyProvider.removeKey(keyId);
           } else {
             this.logger.warn('Key was not removed. Still in use');
-            this.dismiss();
           }
         }
+        setTimeout(() => {
+          this.close();
+        }, 1000);
       })
       .catch(err => {
         this.onGoingProcessProvider.clear();
         let errorText = this.translate.instant('Error');
         this.popupProvider.ionicAlert(errorText, err.message || err);
       });
-  }
-
-  public dismiss() {
-    this.events.publish('Local/WalletListChange');
-    setTimeout(() => {
-      this.viewCtrl.dismiss();
-    }, 1000);
-  }
-
-  public showFullInfo(): void {
-    const infoSheet = this.actionSheetProvider.createInfoSheet('copayers', {
-      secret: this.secret
-    });
-    infoSheet.present();
   }
 
   public shareAddress(): void {
